Disable unnecessary repositories
We frequently see the bugwatcher job fail to build the container image
due to the inability to fetch these repositories.

Disable them as they're not needed.

The usual way to disable the repositories with DNF would be to use the
`dnf config-manager --set-disabled` command, however it is part of
dnf-plugins-core package that is not installed in the fedora container
image. Instead we modify the repository files directly with `sed`.