pretriage: use tenacity to reach Bugzilla
Bugzilla is not always stable, so let's wrap the calls into a function
and use tenacity to retry the calls if Bugzilla isn't working at that
time. We'll try 10 times, with a break of 5 seconds between retries.

This will reduce the number of issues in the job that takes care of
pretriaging bugs for us.