doctext: Fetch all target releases
Prior to this patch, the query for `doctext` looked for verified and
closed bugs targeting 4.10. With this change, the query does not check
the target version any more; instead, it only returns bugs in VERIFIED
or ON_QA.

Note that with this patch, the bugs in CLOSED state are no longer in
scope.