doctext: Also consider Release Note Type
As noted previously, issues.redhat.com is using a new custom field,
'customfield_12317313', for 'Release Note Text'. However, it's also
using a wholly new custom field, 'customfield_12320850' or 'Release Note
Type'. This can be set to various values, one of which is 'No Doc
Update'. If set to this value, clearly we can get away with no setting
the 'Release Note Text' field.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>