Simplify bug query
To track all pertinent bugs, our Jira board historically has two parts:
* include all OCPBUGS issues with our components
* include all OCPBUGS with components adjacent to ours, and with
  "openstack" in the summary

In Bugzilla, this was useful.

However, the second part of the query only seems to catch false
positives lately.

This commit removes the second part, simplifying the bug query to just
trust the Component assignment of the bug.